Move ./js/* to ./src/js/ and ./css/* to ./public/
