+++ src/js/alertBanner.js
